Extend PlatformSpecificTools CommonCpp api with an API to retrieve notifications olm account storage path
Summary: This differential extends  PlatformSpecificTools CommonCpp api with an API to retrieve notifications olm account storage path. It also implements it on iOS and in C++ on Android. Relevant native Java method is implemented in parent differential to this one and is called via JNI
in Android C++ implementation.

Test Plan: Call `PlatformSpecificTools::getNotificationsCryptoAccountPath()` in NotificationService.mm and in AppDelegate.mm. This tests iOS. Call it somewhere in SQLiteQueryEXecutor or CommCoreModule. This tests on Android (and iOS as well).

Reviewers: bartek, tomek, kamil

Subscribers: