[web] Move web/apps directory to web/modals/apps
Summary:
issue: https://linear.app/comm/issue/ENG-3044/change-apps-tab-to-be-a-modal
Moving files that correspond to the Apps tab, because according to the new designs the tab is supposed to be changed into a modal. But not that much of the current code will be changed, so to
make the review easier I'm moving the code first, an will change it later.

Test Plan:
Run web app, checked that pressing the "+" button in  the top bar still opoent the Apps tab, and it displays and works properly.

Reviewers:
kamil, bartek, kuba

Subscribers: