Implement public method to initialize (generate and persist new) notifications olm account
Summary:
This differential implements public method to initialize (generate if not exists and persist new) notifications olm account. Account generation follows the same semantics as public olm account generation in CommCoreModule. If the account already exists it will ne be re-created.
Atomic and fault-tolerant implementation of private `serializeAndFlush` method guarantees that if the file exists it is created correctly.

Test Plan: Call this method in CommCoreModule.

Reviewers: tomek, kamil, bartek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6777