[web] Change apps tab to be a modal
Summary:
issue: https://linear.app/comm/issue/ENG-3044/change-apps-tab-to-be-a-modal
The Apps tab is changed to be a modal. Mind that the designs have changed a bit, see comments under the issue.

Test Plan:
Run web app, pressed the "+" button in the top bar, checked that a modal appears. Check that the modal displays correclty and it is possible to add / remove calendar as an app (the modal works
correctly)

Reviewers:
kamil, kuba, michal

Subscribers: