Refactor CommCoreModule.initializeCryptoAccount not to take userID as an argument.
Summary:
This differential removes userID argument from initializeCryptoAccount method in CommCoreModule since it it not used and is not necessary.
I understand that this differential covers a lot of files, but two of them are just effect of running `yarn codegen-jsi` and I think ot would be hard to split the rest without creating a differential that breaks CI or introduces changes that make app temporarily not working correctly.

Test Plan: Build the app. Ensure public keys are correctly created/retrieved in places where this method of CommCoreModule is called in JS.

Reviewers: atul, bartek, tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6780