[lib][native][web] Set sidebarCreation flag when creating messages that triggered a sidebar
Summary:
This is the final diff in the stack. We set `sidebarCreation` on the client, which makes it so notifs aren't sent by the keyserver.

Depends on D6799

Test Plan:
I tested this whole stack as follows:

1. Create a sidebar by sending a text message on web
2. Create a sidebar by sending a media message on web
3. Create a sidebar by sending a text message on native
4. Create a sidebar by sending a media message on native

In all of these cases, I confirmed that only one notif was delivered (the notif for the text/media message was skipped)

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6800