[keyserver] Removing  from the rust-node-addon gitignore and populate it
Summary:
This diff removes `Cargo.lock` file from the `.gitignore` in the rust-node-addon and populates this file. We should add a lock file to lock the dependencies versions and avoid possible errors due to the versions inconsistency.

Linear task: [[ https://linear.app/comm/issue/ENG-3048/getting-rust-node-addon-error-when-running-yarn-cleaninstall | ENG-3048 ]]

Test Plan: Running `yarn cleaninstall` from the root of the project results in a successful build of the `rust-node-addon` library as a part of the dependencies.

Reviewers: ginsu, varun, jon, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6791