[web] Change import path and fix errors showing up when the typing is correct
Summary:
As mentioned in D6812 the import has to be fixed. We use `useSelector` from web/redux/redux-utils.js on web, because it adds proper typing. This change broke this file, because we should not be
quering an object for a field it doesn't have. Thus the other changes.

Test Plan:
Run web app, checked that Apps directory modal displays its items correclty.

Reviewers:
michal, tomek, kamil
Subscribers: