[web] Register service worker for notifs
Summary: We need to register a service worker to get push notifs from the browser. This diff registers the service worker from the endpoint exposed by the keyserver in one of the previous diffs on page load. For now this service worker only logs when installing and activating.

Test Plan:
- check if service worker is correctly registered (you can check that in the browser's dev tools)
- check that the service worker is correctly updated after changing the service worker file

Reviewers: tomek, atul, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6818