[web] Add notification permission modal
Summary: Adds a simple modal that asks for permission to display notifications. If user hasn't blocked the notifs yet we will request permission from the browser. If we get it we will then create a new subscription. The modal will be used in the laters diffs.

Test Plan:
Check that the user is correctly asked for a permission and the subscription is correctly created.
{F400277}

Reviewers: ashoat, tomek

Reviewed By: ashoat

Subscribers: atul

Differential Revision: https://phab.comm.dev/D6820