[web] Allow users to enable push notifs
Summary: Depending on notification permission we either want to show a modal that asks for permission or (if we already have the permission) create a notification subscription immediately. We want to do this both for existing users and when someone new logs in.

Test Plan: Test if subscription is created/ the modal shows up when someone logs in or open the page when logged in

Reviewers: tomek, kamil, inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6833