[lib] Add invariant to SidebarSource notificationTexts
Summary: While we //do// indicate `SidebarSource` generates notifs in `generatesNotifs`, it should **always** be collapsed, and should never appear as the most recent message for a given collapse key (since `CreateSidebar` is created at the same time and always is considered "newer").

To make sure that `notificationTexts` is never called for `SidebarSource`, I'm adding this `invariant` here.

Depends on D6869

Test Plan: Made sure this `invariant` never gets triggered in a variety of sidebar creation scenarios

Reviewers: atul, ginsu

Subscribers: