[lib] Allow multiple messages in CreateSidebar notificationTexts
Summary:
In a future diff we introduce a `notificationCollapseKey` so that we have access to the `SIDEBAR_SOURCE` when generating the notif text for the `CREATE_SIDEBAR` mention.

Ahead of that, this diff updates the `notificationTexts` code so that it doesn't assert that there is only one message in the collapse result.

Depends on D6867

Test Plan: In combination with the rest of the stack, I confirmed that a sidebar started in response to a text message that @-mentions somebody will result in the @-mentioned user receiving a custom notif that ends with "that tagged you"

Reviewers: atul, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6868