[lib] Collapse CreateSidebar and SidebarSource notifs together
Summary:
Before this diff, `SidebarSource` did not generate any notifs, and `CreateSidebar` did.

After this diff, technically `SidebarSource` generates a notif... but not really. Since `SidebarSource` and `CreateSidebar` get collapsed into each other, and `CreateSidebar` is the more recent message, AND the two messages are **always** generated together on the `keyserver`... the result is that they only ever generate one notif for the both of them.

The only reason for this is that I want to have access to the `SidebarSource` message in `CreateSidebar`'s `notificationTexts` function so that I can check it for @-mentions.

Depends on D6868

Test Plan: In combination with the rest of the stack, I confirmed that a sidebar started in response to a text message that @-mentions somebody will result in the @-mentioned user receiving a custom notif that ends with "that tagged you"

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6869