[keyserver] update identity service client to add authorization metadata
Summary: the identity service is now set up to only accept authorized requests. the keyserver needs to provide an agreed upon token as part of the request metadata.

Test Plan: called the registerUser RPC from keyserver, and it only worked with the correct authorization token in the request metadata.

i.e. if the wrong token is set as an environment variable, an error status is returned

Reviewers: jon, tomek, bartek

Subscribers: