[web] Change "Sign in" button colors to match designs
Summary:
We reverted the change that replaced the purple globally and opted for just changing it in `SIWELoginForm` and `TraditionalLoginForm` to match designs.

Context: We should be able to close the "SIWE on web" design task (https://linear.app/comm/issue/DES-23/siwe-on-web)

Test Plan: Looks as expected, can go through with David/Ted to make sure things look as expected.

Reviewers: ashoat, tomek, ginsu, rohan

Differential Revision: https://phab.comm.dev/D6828