[services][feature-flags] Determine enabled features set
Summary:
Determine a set of enabled features based on the provided params. The idea is described in https://linear.app/comm/issue/ENG-2614/specify-requirements-of-feature-flags-service but the most important part is that when determining a config we take it from the newest version not younger that the version we check.

Depends on D6857

Test Plan: Called the function and checked if it returns correct value depending on the params. It is though a good idea to have some unit tests for it.

Reviewers: michal, bartek, jon, varun

Reviewed By: bartek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6858