[desktop] Allow building with a provisioning profile
Summary: We need aditional entitlements for push notifs on macos. Because of that we also need a provisioning profile to grant us these entitlements. This diff modifies the configuration that detects if there's a provisoning profile in the `desktop` dir and if they exist it uses them and the correct entitlements. The names of the files for the provisoning profiles are just the default files names you get when downloading them from apple website.

Test Plan: Check if it build correctly `yarn package-dev` with no and dev provisioning profile. Check if `yarn package` builds correctly with the production provisioning profile.

Reviewers: tomek, atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6887