[keyserver] Introduce isWebPlatform function
Summary: Step towards handling desktop device types, introduces `isWebPlatform` function that for now only checks if the platform is `web` but in the later diffs it will be expanded to also check for `windows` and `macos`. This diff shouldn't introduce any logic changes.

Test Plan: Search for all occurences of `'web'` and `platform` that should be changed, run flow.

Reviewers: tomek, inka, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6888