[services][feature-flags] Add hello world http server
Summary:
Just a hello world http server. There are a couple of interesting servers available in Rust, but I've chosen this because it is relatively high level, performant, and easily integrates with https.

Depends on D6859

Test Plan: Send GET to `/` and check if response is returned. Checked if both `127.0.0.1` and `[::1]` work.

Reviewers: michal, bartek, jon, varun

Reviewed By: bartek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6860