[keyserver/web/desktop] Add macos and windows as platforms
Summary: We want to support desktop device types to be able to handle native notifications. This diff adds `macos` and `windows` to the possible platform and sets config on the web app depending on where it's running. Older desktop clients with newer web app will still be treated as `web`. Additionaly if on the socket initialization we detect that someone has  a user agent with "Electron" and still has device token of `web` we will ask the web app to send us the current device type, so already logged in users will be updated.

Test Plan:
- check that after login the device type is `macos`
- check that for already logged in users the device type changes to `macos` and remains that after logging out on the anoonymous cookie

Reviewers: tomek, inka, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6889