[keyserver] Make sure we fetch username of all notif recipients
Summary:
This honestly probably isn't necessary, since we already fetch usernames for all thread members, and you shouldn't be receiving a notif unless you're a thread member.

But it doesn't hurt to add this to the set here, and there might be a scenario in the future where somebody is able to receive a notif without being a thread member.

Depends on D6870

Test Plan: There's not much to test here, to be honest... I actually think this diff is a no-op in the current scenario. It's more about future-proofing. We should always fetch the recipient's username

Reviewers: atul, ginsu

Subscribers: