[keyserver] Update notif prepare functions to take bag of params
Summary:
These functions take a long list of inputs. I'm about to add one more, so I figured it would be best to refactor them to take an object of params instead of a list.

Depends on D6871

Test Plan: Flow. Also in combination with the rest of the stack, I confirmed that a sidebar started in response to a text message that @-mentions somebody will result in the @-mentioned user receiving a custom notif that ends with "that tagged you"

Reviewers: atul, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6872