[lib] Make CreateSidebar notifs say when the OP tags the notif recipient
Summary:
Typically you would only receive a `CreateSidebar` notif if you are the author of the OP. However, some earlier diffs introduced a new scenario: if you are tagged in the OP, you will be auto-added to the sidebar at creation time, and as such will receive the `CreateSidebar` notif.

To make sure the notif recipient isn't confused by the notif, this diff adds some text to the notification to clarify this.

Depends on D6873

Test Plan: I confirmed that a sidebar started in response to a text message that @-mentions somebody will result in the @-mentioned user receiving a custom notif that ends with "that tagged you"

Reviewers: atul, ginsu

Subscribers: