[services] Tunnelbroker - Disabling checking of the persisted messages on stream connect
Summary:
This diff introduces changes to the Tunnelbroker gRPC messages stream implementation according to the [[ https://linear.app/comm/issue/ENG-2641/disabling-messages-persistence-in-the-database-in-tunnelbroker | ENG-2641 ]] task, where we need to disable messages persistence in the DynamoDB database using the config file parameter from D6622.
In this diff, we are disabling the check for new messages in the database when the client connects to the stream if the config flag (D6622) is presented.

Linear task: [[ https://linear.app/comm/issue/ENG-2641/disabling-messages-persistence-in-the-database-in-tunnelbroker | ENG-2641 ]]

Test Plan:
1. CI gates are passed.
2. Manual testing:
 The client connects to the gRPC messages stream when messages exist for a certain `deviceID` in the DynamoDB database. When the `messages.skip_persistence` config flag has presented the messages from the database will not be delivered to the stream, because we are skipping looks into the database for them.

Reviewers: jon, bartek

Reviewed By: jon

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6623