[services] Tunnelbroker - Disabling messages persistence on messages removing
Summary:
This diff introduces changes to the Rust gRPC messages stream handler to disable removing messages from the DynamoDB database when the config flag from the D6622 is presented. The reason to skip messages removal is the [[ https://linear.app/comm/issue/ENG-2641/disabling-messages-persistence-in-the-database-in-tunnelbroker | ENG-2641 ]] task where we are skipping the persistence of the message in the database in the case when the config flag is presented.

Linear task: [[ https://linear.app/comm/issue/ENG-2641/disabling-messages-persistence-in-the-database-in-tunnelbroker | ENG-2641 ]]

Test Plan:
1. Passing the CI gates.
2. Manual testing:
 Connect to the gRPC messages stream and send the message to the certain `deviceID`. Connect to the stream as a client on this `deviceID` to which the message was sent. Consume the message. The expected result is that the message will not be persisted in the DynamoDB messages table and there is no error when trying to remove the message on consuming.

Reviewers: jon, bartek

Reviewed By: jon, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6626