[lib/native/web] switch reactions data structure from map to plain object
Summary: In dev mode the app was giving a warning whenever a user presses a message that "Non-serializble values were found in the navigation state". This was caused becasue reactions were stored in a map. To fix this I switched the data structure from a map to a plain object

Test Plan: When a user taps on a message now the warning message about "non-serializable..." no longer shows up

Reviewers:

Subscribers: