[lib/native/web] switch reactions data structure from map to plain object
Summary:
In dev mode the app was warning whenever a user pressed a message that "Non-serializble values were found in the navigation state". This was caused becasue reactions were stored in a map. To fix this, I switched the data structure from a map to a plain object

---

Linear Task: [[ https://linear.app/comm/issue/ENG-2851/reactions-list-in-textmessagetooltipmodal-is-persisting-non-json | ENG-2851 ]]

Test Plan:
When a user taps on a message now the warning message about "non-serializable..." no longer shows up

Before:
{F399480}

After:
{F399481}

In addition tested on both `native` and `web` the following to make sure there were no regressions with reactions:
- reacting to a message (text/multimedia/robotext)
- unreacting to a message (text/multimedia/robotext)
- viewing who has liked a message
- sending multiple reactions to the same message

also `flow`

Reviewers: atul, tomek, ashoat, rohan

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6875