[web] Add `PickledOLMAccount`s to Redux `CryptoStore`
Summary:
We need to pickle (ie serialize) the OLM accounts so we can eg sign messages. In this diff we add the `PickedOLMAccount`s to the Redux `CryptoStore` type.

NOTE: We aren't actually populating the fields yet, just adding them.

---

Depends on D6881

Test Plan: This should be a noop.

Reviewers: ashoat, tomek, marcin

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6882