[web] "Unpickle" primary OLM account to sign keys and get `SignedIdentityKeysBlob`
Summary:
We unpickle the primary OLM account in order to sign a "concatenation" of the primary and notification keys.

We then bundle up the `payload` and `signature` into `SignedIdentityKeysBlob` which is what will get included in the `login` request to the `keyserver`.

---

Depends on D6883

Test Plan:
Logged `identity_keys()` of "unpickled" account and it matched the keys from the instantiation of the primary account:

{F400019}

`signedIdentityKeysBlob` is created as expected:
{F400017}

Reviewers: ashoat, tomek, marcin

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6884