[services][feature-flags] Dockerize the service
Summary:
Introduce Docker config so that the service can be run in a container.

Depends on D6862

Test Plan: Run the container locally. Using Docker Desktop terminal send a GET to the service and check if it returns a response based on AWS DynamoDB content. Haven't figured out how to send a request from host, but verifying that is the next step.

Reviewers: michal, bartek, jon, varun

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6886