[keyserver] Extend ios notifs to handle macos
Summary:
We can reuse most of out already existing code for macOS notifs. This diff mostly just replaces `codeVersion` with full `platformDetails` so we can pass a platform and change for example the apn topic depending if we are sending the notifcation to the ios or macos.

Depends on D6891

Test Plan: Check if notifications still work on ios

Reviewers: tomek, inka, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6892