[keyserver] Send macos notifications
Summary: Actually send the notification to all macos devices. One things that is different is that we don't include text for `alert` notif so it doesn't display automatically by macos but we handle the visible notification ourselves. More information in [ENG-3056](https://linear.app/comm/issue/ENG-3056).

Test Plan: Tested with the next diff, if the notifications are sent correctly.

Reviewers: tomek, inka, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6893