[web] Pass `signedIdentityKeysBlob` to `keyserver` on `siwe_auth`
Summary:
Include `signedIdentityKeysBlob` in `siwe_auth` request and update `siweAuthRequestInputValidator` to accept `signedIdentityKeysBlob`. Right now we don't do anything with the `signedIdentityKeysBlob`, but it's being passed to `keyserver` from `web` when hitting both `login` and `siwe_auth` endpoints.

Next up is
1. Making sure that `signedIdentityKeysBlob` is valid (regex checks)
2. Pulling out `primary.ed25519` and verifying the `payload` signature
(maybe an argument for leaving `primaryIdentityPublicKey` as part of the `login` request and maybe adding it to `siwe_auth`?)
3. Stringifying the entire `signedIdentityKeysBlob` and storing it in the serverDB (going to do a migration to remove existing stuff first).

---

Depends on D6895

Test Plan:
Set breakpoint in `siweAuthResponder` and observed that `signedIdentityKeysBlob` "made it through" and validator successfully validated request:

Reviewers:
ashoat, tomek