[DRAFT] Verify `signedIdentityKeysBlob` signature in `logInResponder`
Summary:
Pass `primaryIdentityPublicKeys.ed25519`, `signedIdentityKeysBlob.payload` and `signedIdentityKeysBlob.signature` to `olmUtil.ed25519_verify(...)` to ensure that the `signature` is valid.

The reason this diff is titled `[DRAFT]` is because of the way we're calling `olm.init()` and `new olm.Utility()` every time `logInResponder` encounters a request with `signedIdentityKeysBlob`. Instead we'll want to probably run `olm.init()` once when `keyserver` starts up, and we'll want to instantiate `olm.Utility()` once and make it accessible globally.

---

Depends on D6896

Test Plan:
I tested with a valid signature and observed that `ed25519_verify(...)` succeeded. I tested with an invalid signature ("blah"), and observed that `olmUtil.ed25519(...)` threw an exception.

Reviewers:
ashoat, tomek