[keyserver] Unifying API interface and adding Tunnelbroker types
Summary:
This diff introduces adding the Tunnelbroker Rust binding types, and refactoring the API getter function to return the API types for all Rust bindings.
Types for the Rust API bindings are moved to a separate file.
The purpose of the getter is to call the Tunnelbroker client class and Identity service API from the nodejs codebase.

Linear task: [[ https://linear.app/comm/issue/ENG-2613/implement-an-abstraction-for-the-redis-and-tunnelbroker-as-a-pub-sub | ENG-2613 ]]

Test Plan: Passing the CI gates seems like enough for these changes.

Reviewers: varun, ashoat, jon

Reviewed By: varun, ashoat

Subscribers: jon, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6690