[services][feature-flags] Introduce map and i32 parsing
Summary:
We have to introduce a function that parses strings into numbers, because map keys in Dynamo are plain strings instead of `AttributeValue`s. This affects what we store in `DBItemError` - now it is either `AttributeValue` or a String.

Depends on D6854

Test Plan: Tested with the rest of the stack.

Reviewers: michal, bartek, jon, varun

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6855