[keyserver] Verify `signedIdentityKeysBlob` signature in `logInResponder`
Summary:
Pass `primaryIdentityPublicKeys.ed25519`, `signedIdentityKeysBlob.payload` and `signedIdentityKeysBlob.signature` to `olmUtil.ed25519_verify(...)` to ensure that the `signature` is valid.

---

Depends on D6896

Test Plan: I tested with a valid signature and observed that `ed25519_verify(...)` succeeded. I tested with an invalid signature ("blah"), and observed that `olmUtil.ed25519(...)` threw an exception.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6897