[keyserver] Only set sidebarCreation flag for the first message in the sidebar
Summary:
When I say "first message" in the commit title, I really mean the third message, as each sidebar starts with a `SIDEBAR_SOURCE` and then a `CREATE_SIDEBAR`.

Check out the code comment for an explanation as to why we're doing this.

Test Plan:
Force `sidebarCreation` to `true` on the client for all text messages, and confirm that it is ignored when invalid

Reviewers: atul

Subscribers: