[identity] add session initialization info to RegisterUser and LoginUser RPCs
Summary:
We need to send session initialization info to the identity service as part of the registration and login workflows so that the service can provide this info to clients trying to establish a new channel of communication subsequently with a given user.

The contents of the session initialization map are described in the .proto file.

Test Plan:
1. Tried to register a new user without session initialization info, failed as expected
2. Tried to register a new user with session initialization info, succeeded and DDB showed the keys + signature + social proof
3. Tried to log in an existing user with new session initialization info (i.e. different device), succeeded and DDB showed both sets of info, keyed by the public key

Reviewers: tomek, bartek, jon

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6910