[keyserver] Check `primaryIdentityPublicKeys.ed25519` against SIWE statement `primaryIdentityPublicKey`
Summary:
We want to make sure that the public key included in the SIWE statement matches that provided in the `siwe_auth` request.

Test Plan:
SIWE on web and set breakpoints and things were as expected.

Going to test `native` before landing to make sure `siwe_auth` continues working as expected.

Reviewers:
ashoat, tomek