[keyserver][lib] Allow notificationTexts to return null
Summary:
In order to decide whether to send a notif, we may need more information than we have at the time `generatesNotifs` is called.

More concretely: in a later diff, we'll need to make a decision on whether to send a notif based on the contents of an earlier message with the same collapse key.

Depends on D6935

Test Plan:
Flow, and tested in combination with the next diffs to confirm that getting @-tagged in a message that triggers the creation of a sidebar will update the `SIDEBAR_CREATION` notif
to be suffixed with "and tagged you"

Reviewers: atul